+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v0.2.0 *(2021-01-24)*
+
+- Add an option not to transfer archived cards (#2)
+
+- Fix GitLab users fetching method (#4)
+- Fix closed issue detection based on closed card and list (#3)
+
 ## v0.1.0 *(2020-11-19)*
 
 - Initial release